refactor: move getListOfStageFiles from Merger to Sorter
